Player controller fix and template level
I've simplified the player controller, so that it's easier to work with.
I found the old one I write still had some bugs. I've also created a
template level so we can stay on the same page as far as level sizing
goes. The template level uses a sphere as the main character, as we
still don't have assets yet.